Move iqvia base entity to separate module
