+++ homeassistant/components/iqvia/sensor.py
@@ -17,7 +17,6 @@ from homeassistant.const import ATTR_STATE
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import IQVIAEntity
 from .const import (
     DOMAIN,
     TYPE_ALLERGY_FORECAST,
@@ -33,6 +32,7 @@ from .const import (
     TYPE_DISEASE_INDEX,
     TYPE_DISEASE_TODAY,
 )
+from .entity import IQVIAEntity
 
 ATTR_ALLERGEN_AMOUNT = "allergen_amount"
 ATTR_ALLERGEN_GENUS = "allergen_genus"
